Refactor: migrate landing page to Vite
Move the landing page assets into a Vite-powered modular structure.
Update the page markup,product buttons, plan section, and event
handling for the new source layout.

Also prepare project documentation and ignore generated dependencies/build
artifacts.
