+++ README.md
@@ -1,0 +1,45 @@
+# Hawas Coffee
+
+A responsive landing page for Hawas Coffee, built with Vite, vanilla JavaScript, and CSS.
+
+## Features
+
+- Hero section with branded product imagery
+- Product selection cards
+- Subscription plan section
+- Responsive layout for desktop and mobile
+
+## Tech Stack
+
+- Vite
+- Vanilla JavaScript
+- CSS
+
+## Getting Started
+
+Install dependencies:
+
+```sh
+pnpm install
+
+Start the development server:
+
+pnpm dev
+
+Create a production build:
+
+pnpm build
+
+Preview the production build:
+
+pnpm preview
+
+## Project Structure
+
+- index.html - page markup
+- src/main.js - app entry point
+- src/listener.js - UI event handling
+- src/store/productData.js - product data
+- src/styles/ - reset, desktop, and mobile styles
+- media/ - images and brand assets
+```
