+++ src/main.js
@@ -1,0 +1,9 @@
+import { addListener } from "./listener.js";
+import "./styles/cssreset.css";
+import "./styles/styles.css";
+import "./styles/mobile.css";
+function pageInit() {
+  addListener();
+}
+
+pageInit();
