+++ src/store/productData.js
@@ -1,0 +1,26 @@
+export const productData = {
+  1: {
+    title: "Viet Bang",
+    img: "../../media/HawasProduct.webp",
+    description:
+      "Single origin Arabica. Medium roast with hints of dark chocolate and pepper. ",
+  },
+  2: {
+    title: "Kenya Bold",
+    img: "../../media/HawasProduct.webp",
+    description:
+      "Strong and full-bodied. Notes of berry and citrus. Perfect for espresso lovers.",
+  },
+  3: {
+    title: "Colombia Light",
+    img: "../../media/HawasProduct.webp",
+    description:
+      "Smooth and mellow. Light roast with caramel and nutty aromas.",
+  },
+  4: {
+    title: "Colombia Tight",
+    img: "../../media/HawasProduct.webp",
+    description:
+      "Smooth and mellow. Light roast with caramel and nutty aromas.",
+  },
+};
